Move RecentFileCache up a level
