+++ src/worker/recent_file_cache.cpp
